Unify C64 startup acknowledgment dialog (audio unlock + ROM consent)
Replace the pre-Avalonia HTML audio-autoplay gate with a single themed
acknowledgment dialog shown before system selection for URL-driven C64 starts:
shows what will run and, when ROMs are missing, the ROM-license consent
(checkbox). Confirming unlocks browser audio autoplay; cancelling aborts before
selection and falls back to the default system.

- Add pre-selection participant seam AcknowledgeStartupAsync + UnlockAudio context
- Add IHostApp.IsSystemConfigValid for selection-free ROM detection
- Refactor C64RomPromptService into C64AcknowledgmentService (also serves the
  config dialog's ROM-only consent, now with checkbox)
- Re-validate config and refresh emulator display size after a host-config update
  so the post-ROM-download UI isn't stale; return to Information tab once resolved
